+++ src/lib/gpu/gpu-splat-rasterizer.ts
@@ -3,7 +3,6 @@ import {
     BUFFERUSAGE_COPY_SRC,
     SHADERLANGUAGE_WGSL,
     SHADERSTAGE_COMPUTE,
-    UNIFORMTYPE_FLOAT,
     UNIFORMTYPE_UINT,
     BindGroupFormat,
     BindStorageBufferFormat,
@@ -17,33 +16,31 @@ import {
     UniformFormat
 } from 'playcanvas';
 
+import { constantsChunk } from './shaders/chunks/constants';
+import { covariance3D } from './shaders/chunks/covariance-3d';
+import { jacobianEquirect } from './shaders/chunks/jacobian-equirect';
+import { jacobianPinhole } from './shaders/chunks/jacobian-pinhole';
+import { projectionEquirect } from './shaders/chunks/projection-equirect';
+import { projectionPinhole } from './shaders/chunks/projection-pinhole';
+import { quatRotation } from './shaders/chunks/quat-rotation';
+import { shBand1 } from './shaders/chunks/sh-band-1';
+import { shBand2 } from './shaders/chunks/sh-band-2';
+import { shBand3 } from './shaders/chunks/sh-band-3';
+import { tileAabbEquirect } from './shaders/chunks/tile-aabb-equirect';
+import { tileAabbPinhole } from './shaders/chunks/tile-aabb-pinhole';
+import { tileWalkEquirect } from './shaders/chunks/tile-walk-equirect';
+import { tileWalkPinhole } from './shaders/chunks/tile-walk-pinhole';
+import { finalizeWgsl } from './shaders/finalize';
+import { findBoundariesWgsl } from './shaders/find-boundaries';
+import { initTileOffsetsWgsl } from './shaders/init-tile-offsets';
+import { prefixSumWgsl } from './shaders/prefix-sum';
+import { prepareIndirectWgsl } from './shaders/prepare-indirect';
+import { projectWgsl } from './shaders/project';
+import { rasterizeBinnedWgsl } from './shaders/rasterize-binned';
+import { tileBinEmitPairsWgsl } from './shaders/tile-bin-emit-pairs';
+import { uniformsStruct, uniformFormatEntries } from './shaders/uniforms';
 import { type Projection } from '../render/camera';
-import {
-    AA_DILATION_COV,
-    DISCRIMINANT_FLOOR,
-    JACOBIAN_LIMIT_FACTOR,
-    MIN_ALPHA,
-    MIN_TRANSMITTANCE,
-    OPACITY_CAP,
-    POLE_EPS,
-    RADIUS_FADE_END_FRAC,
-    RADIUS_FADE_START_FRAC,
-    SIGMA_CUTOFF,
-    TILE_SIZE
-} from '../render/config';
-
-/**
- * Format a JS number as a WGSL `f32` literal. Adds an explicit `.0` so
- * integer-valued constants like `3` aren't parsed as `AbstractInt` —
- * keeps shaders readable when the constant flips to a fractional value.
- *
- * @param n - Numeric value to format.
- * @returns WGSL literal string with explicit `.0` for integer values.
- */
-const wgslF32 = (n: number): string => {
-    const s = n.toString();
-    return s.includes('.') || s.includes('e') || s.includes('E') ? s : `${s}.0`;
-};
+import { TILE_SIZE } from '../render/config';
 
 /** 12 floats per projected splat: vec4 × 3. */
 const PROJECTION_STRIDE_F32 = 12;
@@ -122,950 +119,6 @@ const numSHCoeffsPerChannel = (bands: number): number => {
     return bands === 0 ? 0 : bands === 1 ? 3 : bands === 2 ? 8 : 15;
 };
 
-/**
- * WGSL source for the project compute. Reads raw splat fields from the
- * per-slot input buffer, evaluates SH for view-dependent color, computes
- * 2D inverse covariance + screen-space 3σ radius, and writes a packed
- * projection record per gaussian. Invalid splats (behind near, degenerate
- * covariance, outside group AABB) are written with `radius = 0` so the
- * rasterizer can early-out on the first vec4 load.
- *
- * @param coeffsPerChannel - Per-channel SH coefficient count (0/3/8/15).
- * @param maxCoveragePerSplat - Hard upper bound on per-splat tile count.
- * Embedded into the shader's `coverage[i] = min(raw, cap)` clamp. The
- * orchestrator sets this to the group's full tile area so the clamp is
- * geometrically unreachable in practice; emit-pairs walks the bbox
- * row-major and truncates at the bottom-right if the cap is ever hit.
- * @param projection - Camera projection. `pinhole` swaps in the
- * classical perspective + EWA path; `equirect` swaps in spherical
- * mapping, a non-linear Jacobian, and wrap-aware X tile coverage.
- * @returns WGSL source for the project compute shader.
- */
-const projectWgsl = (coeffsPerChannel: number, maxCoveragePerSplat: number, projection: Projection) => /* wgsl */`
-struct Uniforms {
-    rightX: f32, rightY: f32, rightZ: f32, _p0: f32,
-    downX: f32, downY: f32, downZ: f32, _p1: f32,
-    forwardX: f32, forwardY: f32, forwardZ: f32, _p2: f32,
-    eyeX: f32, eyeY: f32, eyeZ: f32, _p3: f32,
-    focalX: f32, focalY: f32, near: f32, _p4: f32,
-    imageWidth: u32, imageHeight: u32, splatStride: u32, chunkSize: u32,
-    groupPixelMinX: u32, groupPixelMinY: u32, groupPixelMaxX: u32, groupPixelMaxY: u32,
-    groupTilesX: u32, groupTilesY: u32, groupPixelOriginX: u32, groupPixelOriginY: u32,
-    bgR: f32, bgG: f32, bgB: f32, bgA: f32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: Uniforms;
-@group(0) @binding(1) var<storage, read> splats: array<f32>;
-@group(0) @binding(2) var<storage, read_write> projected: array<vec4<f32>>;
-@group(0) @binding(3) var<storage, read_write> coverage: array<u32>;
-
-const SH_C0: f32 = 0.28209479177387814;
-const SH_C1: f32 = 0.4886025119029199;
-const SH_C2_0: f32 = 1.0925484305920792;
-const SH_C2_1: f32 = -1.0925484305920792;
-const SH_C2_2: f32 = 0.31539156525252005;
-const SH_C2_3: f32 = -1.0925484305920792;
-const SH_C2_4: f32 = 0.5462742152960396;
-const SH_C3_0: f32 = -0.5900435899266435;
-const SH_C3_1: f32 = 2.890611442640554;
-const SH_C3_2: f32 = -0.4570457994644658;
-const SH_C3_3: f32 = 0.3731763325901154;
-const SH_C3_4: f32 = -0.4570457994644658;
-const SH_C3_5: f32 = 1.445305721320277;
-const SH_C3_6: f32 = -0.5900435899266435;
-
-const COEFFS_PER_CHANNEL: u32 = ${coeffsPerChannel}u;
-
-fn writeInvalid(idx: u32) {
-    projected[idx * 3u + 0u] = vec4<f32>(0.0, 0.0, 0.0, 0.0);
-    projected[idx * 3u + 1u] = vec4<f32>(0.0, 0.0, 0.0, 0.0);
-    projected[idx * 3u + 2u] = vec4<f32>(0.0, 0.0, 0.0, 0.0);
-    coverage[idx] = 0u;
-}
-
-@compute @workgroup_size(64)
-fn main(@builtin(global_invocation_id) gid: vec3<u32>) {
-    let i = gid.x;
-    if (i >= uniforms.chunkSize) { return; }
-
-    let base = i * uniforms.splatStride;
-
-    let posX = splats[base + 0u];
-    let posY = splats[base + 1u];
-    let posZ = splats[base + 2u];
-    let rotW = splats[base + 3u];
-    let rotX = splats[base + 4u];
-    let rotY = splats[base + 5u];
-    let rotZ = splats[base + 6u];
-    let lsX = splats[base + 7u];
-    let lsY = splats[base + 8u];
-    let lsZ = splats[base + 9u];
-    let opacity = splats[base + 10u];
-    let fdcR = splats[base + 11u];
-    let fdcG = splats[base + 12u];
-    let fdcB = splats[base + 13u];
-
-    // World → camera
-    let wx = posX - uniforms.eyeX;
-    let wy = posY - uniforms.eyeY;
-    let wz = posZ - uniforms.eyeZ;
-    let cx = uniforms.rightX * wx + uniforms.rightY * wy + uniforms.rightZ * wz;
-    let cy = uniforms.downX * wx + uniforms.downY * wy + uniforms.downZ * wz;
-    let cz = uniforms.forwardX * wx + uniforms.forwardY * wy + uniforms.forwardZ * wz;
-
-${projection === 'pinhole' ? /* wgsl */`
-    if (cz <= uniforms.near) { writeInvalid(i); return; }
-
-    let invZ = 1.0 / cz;
-    let screenX = uniforms.focalX * cx * invZ + f32(uniforms.imageWidth) * 0.5;
-    let screenY = uniforms.focalY * cy * invZ + f32(uniforms.imageHeight) * 0.5;
-` : /* wgsl */`
-    // Equirect: cull by radial distance (splats inside the near sphere
-    // would otherwise project to a wildly unstable longitude). The
-    // longitude is undefined at the camera origin and degenerates near
-    // the poles; the POLE_EPS clamp on rxz below keeps the Jacobian
-    // bounded for splats arbitrarily close to the zenith/nadir.
-    let r2 = cx * cx + cy * cy + cz * cz;
-    if (r2 <= uniforms.near * uniforms.near) { writeInvalid(i); return; }
-    let r = sqrt(r2);
-    let rxz2 = cx * cx + cz * cz;
-    let rxz = sqrt(rxz2);
-    let rxzClamped = max(rxz, ${wgslF32(POLE_EPS)} * r);
-
-    let invTwoPi: f32 = 0.15915494309189535;
-    let invPi:    f32 = 0.3183098861837907;
-    let imgWf = f32(uniforms.imageWidth);
-    let imgHf = f32(uniforms.imageHeight);
-    let lon = atan2(cx, cz);
-    let sinLat = clamp(cy / r, -1.0, 1.0);
-    let lat = asin(sinLat);
-    let screenX = (lon * invTwoPi + 0.5) * imgWf;
-    // cy > 0 = camera-down axis = below horizon → lat > 0 → screenY in
-    // the bottom half. screenY = 0 maps to the zenith (above the camera).
-    let screenY = (lat * invPi + 0.5) * imgHf;
-`}
-
-    // Quaternion normalize.
-    let qlen2 = rotW * rotW + rotX * rotX + rotY * rotY + rotZ * rotZ;
-    if (qlen2 == 0.0) { writeInvalid(i); return; }
-    let invQ = inverseSqrt(qlen2);
-    let qw = rotW * invQ;
-    let qx = rotX * invQ;
-    let qy = rotY * invQ;
-    let qz = rotZ * invQ;
-
-    // Rotation matrix from unit quaternion.
-    let xx = qx * qx; let yy = qy * qy; let zz = qz * qz;
-    let xy = qx * qy; let xzq = qx * qz; let yz = qy * qz;
-    let wxq = qw * qx; let wy_ = qw * qy; let wzq = qw * qz;
-    let r00 = 1.0 - 2.0 * (yy + zz);
-    let r01 = 2.0 * (xy - wzq);
-    let r02 = 2.0 * (xzq + wy_);
-    let r10 = 2.0 * (xy + wzq);
-    let r11 = 1.0 - 2.0 * (xx + zz);
-    let r12 = 2.0 * (yz - wxq);
-    let r20 = 2.0 * (xzq - wy_);
-    let r21 = 2.0 * (yz + wxq);
-    let r22 = 1.0 - 2.0 * (xx + yy);
-
-    let sx = exp(lsX);
-    let sy = exp(lsY);
-    let sz = exp(lsZ);
-
-    let m00 = r00 * sx; let m01 = r01 * sy; let m02 = r02 * sz;
-    let m10 = r10 * sx; let m11 = r11 * sy; let m12 = r12 * sz;
-    let m20 = r20 * sx; let m21 = r21 * sy; let m22 = r22 * sz;
-
-    let sig00 = m00 * m00 + m01 * m01 + m02 * m02;
-    let sig01 = m00 * m10 + m01 * m11 + m02 * m12;
-    let sig02 = m00 * m20 + m01 * m21 + m02 * m22;
-    let sig11 = m10 * m10 + m11 * m11 + m12 * m12;
-    let sig12 = m10 * m20 + m11 * m21 + m12 * m22;
-    let sig22 = m20 * m20 + m21 * m21 + m22 * m22;
-
-    let v00 = uniforms.rightX; let v01 = uniforms.rightY; let v02 = uniforms.rightZ;
-    let v10 = uniforms.downX; let v11 = uniforms.downY; let v12 = uniforms.downZ;
-    let v20 = uniforms.forwardX; let v21 = uniforms.forwardY; let v22 = uniforms.forwardZ;
-
-    let t00 = v00 * sig00 + v01 * sig01 + v02 * sig02;
-    let t01 = v00 * sig01 + v01 * sig11 + v02 * sig12;
-    let t02 = v00 * sig02 + v01 * sig12 + v02 * sig22;
-    let t10 = v10 * sig00 + v11 * sig01 + v12 * sig02;
-    let t11 = v10 * sig01 + v11 * sig11 + v12 * sig12;
-    let t12 = v10 * sig02 + v11 * sig12 + v12 * sig22;
-    let t20 = v20 * sig00 + v21 * sig01 + v22 * sig02;
-    let t21 = v20 * sig01 + v21 * sig11 + v22 * sig12;
-    let t22 = v20 * sig02 + v21 * sig12 + v22 * sig22;
-
-    let c00 = t00 * v00 + t01 * v01 + t02 * v02;
-    let c01 = t00 * v10 + t01 * v11 + t02 * v12;
-    let c02 = t00 * v20 + t01 * v21 + t02 * v22;
-    let c11 = t10 * v10 + t11 * v11 + t12 * v12;
-    let c12 = t10 * v20 + t11 * v21 + t12 * v22;
-    let c22 = t20 * v20 + t21 * v21 + t22 * v22;
-
-${projection === 'pinhole' ? /* wgsl */`
-    // Pinhole Jacobian with x/z, y/z clamp. J is the 2×3 matrix
-    //   [[jx0,   0, jx2],
-    //    [  0, jy1, jy2]]
-    // — the zero entries let us drop the u01·jy0 / u10 / u11·jy0 terms
-    // in cov = J·Σ·Jᵀ that the equirect path retains.
-    let limX = ${wgslF32(JACOBIAN_LIMIT_FACTOR)} * (f32(uniforms.imageWidth) * 0.5) / uniforms.focalX;
-    let limY = ${wgslF32(JACOBIAN_LIMIT_FACTOR)} * (f32(uniforms.imageHeight) * 0.5) / uniforms.focalY;
-    let txtz = clamp(cx * invZ, -limX, limX);
-    let tytz = clamp(cy * invZ, -limY, limY);
-    let jcx = txtz * cz;
-    let jcy = tytz * cz;
-    let jx0 = uniforms.focalX * invZ;
-    let jx2 = -uniforms.focalX * jcx * invZ * invZ;
-    let jy1 = uniforms.focalY * invZ;
-    let jy2 = -uniforms.focalY * jcy * invZ * invZ;
-
-    let u00 = jx0 * c00 + jx2 * c02;
-    let u01 = jx0 * c01 + jx2 * c12;
-    let u02 = jx0 * c02 + jx2 * c22;
-    let u11 = jy1 * c11 + jy2 * c12;
-    let u12 = jy1 * c12 + jy2 * c22;
-
-    var cov00 = u00 * jx0 + u02 * jx2;
-    let cov01 = u01 * jy1 + u02 * jy2;
-    var cov11 = u11 * jy1 + u12 * jy2;
-` : /* wgsl */`
-    // Equirect Jacobian. With longitude θ = atan2(cx, cz) and latitude
-    // φ = asin(cy/r) (cy is the camera-down axis, so φ>0 = below the
-    // horizon), the per-axis screen derivatives multiplied by the pixel
-    // scales (kx, ky) = (W/(2π), H/π) give the 2×3 Jacobian:
-    //   ∂screenX/∂(cx,cy,cz) = kx · ( cz/rxz²,  0,             -cx/rxz² )
-    //   ∂screenY/∂(cx,cy,cz) = ky · (-cx·cy/(r²·rxz),  rxz/r²,  -cy·cz/(r²·rxz) )
-    // rxzClamped (≥ POLE_EPS·r) keeps every denominator finite as a
-    // splat approaches the pole. j[0][1] = 0, j[1][0] ≠ 0 — the cov
-    // formula carries the extra u00·jy0 / u10·jy0 / u11·jy0 terms that
-    // the pinhole simplification dropped.
-    let kx = imgWf * invTwoPi;
-    let ky = imgHf * invPi;
-    let invRxzC2 = 1.0 / (rxzClamped * rxzClamped);
-    let invR2 = 1.0 / r2;
-    let invR2Rxz = invR2 / rxzClamped;
-    let jx0 =  kx * cz * invRxzC2;
-    let jx2 = -kx * cx * invRxzC2;
-    let jy0 = -ky * cx * cy * invR2Rxz;
-    let jy1 =  ky * rxzClamped * invR2;
-    let jy2 = -ky * cy * cz * invR2Rxz;
-
-    let u00 = jx0 * c00 + jx2 * c02;
-    let u01 = jx0 * c01 + jx2 * c12;
-    let u02 = jx0 * c02 + jx2 * c22;
-    let u10 = jy0 * c00 + jy1 * c01 + jy2 * c02;
-    let u11 = jy0 * c01 + jy1 * c11 + jy2 * c12;
-    let u12 = jy0 * c02 + jy1 * c12 + jy2 * c22;
-
-    var cov00 = u00 * jx0 + u02 * jx2;
-    let cov01 = u00 * jy0 + u01 * jy1 + u02 * jy2;
-    var cov11 = u10 * jy0 + u11 * jy1 + u12 * jy2;
-`}
-
-    cov00 = cov00 + ${wgslF32(AA_DILATION_COV)};
-    cov11 = cov11 + ${wgslF32(AA_DILATION_COV)};
-
-    let det = cov00 * cov11 - cov01 * cov01;
-    if (det <= 0.0) { writeInvalid(i); return; }
-
-    let invDet = 1.0 / det;
-    let covInvA = cov11 * invDet;
-    let covInvB = -cov01 * invDet;
-    let covInvC = cov00 * invDet;
-
-    let mid = 0.5 * (cov00 + cov11);
-    let disc = sqrt(max(${wgslF32(DISCRIMINANT_FLOOR)}, mid * mid - det));
-    let lambdaMax = mid + disc;
-    let radiusRaw = ${wgslF32(SIGMA_CUTOFF)} * sqrt(lambdaMax);
-
-    // Outlier-splat fade: huge splats (close-by mega-splats or pathological
-    // training output) would otherwise project to a screen-spanning footprint
-    // and tint the whole frame. Linearly fade alpha from 1 to 0 as the
-    // un-clamped radius grows from fadeStart to fadeEnd, and discard
-    // beyond. The bbox we hand to the rasterizer is clamped at fadeEnd
-    // so the binner doesn't reserve tile coverage for a splat that
-    // contributes zero anyway. Softer than a hard clamp: prevents the
-    // visible pop as the camera approaches a clipped splat.
-    //
-    // Thresholds are fractions of image height so the SAME world-space
-    // splats fade at every render resolution — preserves cross-
-    // resolution consistency (e.g. 8K-downsampled-to-1080p matches
-    // 1080p direct).
-    let fadeStart = ${wgslF32(RADIUS_FADE_START_FRAC)} * f32(uniforms.imageHeight);
-    let fadeEnd = ${wgslF32(RADIUS_FADE_END_FRAC)} * f32(uniforms.imageHeight);
-    let radiusFade = clamp((fadeEnd - radiusRaw) / (fadeEnd - fadeStart), 0.0, 1.0);
-    if (radiusFade <= 0.0) { writeInvalid(i); return; }
-    let radius = ceil(min(radiusRaw, fadeEnd));
-
-    // Group AABB cull. The BVH frustum query may include splats whose
-    // 3D AABB grazes the frustum but whose 2D footprint misses the group.
-    let gx0 = f32(uniforms.groupPixelMinX);
-    let gx1 = f32(uniforms.groupPixelMaxX);
-    let gy0 = f32(uniforms.groupPixelMinY);
-    let gy1 = f32(uniforms.groupPixelMaxY);
-    if (screenX + radius < gx0 || screenX - radius >= gx1 ||
-        screenY + radius < gy0 || screenY - radius >= gy1) {
-        writeInvalid(i);
-        return;
-    }
-
-    // View-dependent color via SH evaluation.
-    let dpx = posX - uniforms.eyeX;
-    let dpy = posY - uniforms.eyeY;
-    let dpz = posZ - uniforms.eyeZ;
-    let dirLen = max(1e-30, sqrt(dpx * dpx + dpy * dpy + dpz * dpz));
-    let dirX = dpx / dirLen;
-    let dirY = dpy / dirLen;
-    let dirZ = dpz / dirLen;
-
-    var cR = SH_C0 * fdcR;
-    var cG = SH_C0 * fdcG;
-    var cB = SH_C0 * fdcB;
-
-    ${coeffsPerChannel >= 3 ? `
-    {
-        let n = COEFFS_PER_CHANNEL;
-        let shBase = base + 14u;
-        let b0 = -SH_C1 * dirY;
-        let b1 = SH_C1 * dirZ;
-        let b2 = -SH_C1 * dirX;
-        cR = cR + b0 * splats[shBase + 0u] + b1 * splats[shBase + 1u] + b2 * splats[shBase + 2u];
-        cG = cG + b0 * splats[shBase + n + 0u] + b1 * splats[shBase + n + 1u] + b2 * splats[shBase + n + 2u];
-        cB = cB + b0 * splats[shBase + 2u * n + 0u] + b1 * splats[shBase + 2u * n + 1u] + b2 * splats[shBase + 2u * n + 2u];
-    }` : ''}
-    ${coeffsPerChannel >= 8 ? `
-    {
-        let n = COEFFS_PER_CHANNEL;
-        let shBase = base + 14u;
-        let xx2 = dirX * dirX;
-        let yy2 = dirY * dirY;
-        let zz2 = dirZ * dirZ;
-        let xy2 = dirX * dirY;
-        let yz2 = dirY * dirZ;
-        let xz2 = dirX * dirZ;
-        let b3 = SH_C2_0 * xy2;
-        let b4 = SH_C2_1 * yz2;
-        let b5 = SH_C2_2 * (2.0 * zz2 - xx2 - yy2);
-        let b6 = SH_C2_3 * xz2;
-        let b7 = SH_C2_4 * (xx2 - yy2);
-        cR = cR + b3 * splats[shBase + 3u] + b4 * splats[shBase + 4u] + b5 * splats[shBase + 5u] + b6 * splats[shBase + 6u] + b7 * splats[shBase + 7u];
-        cG = cG + b3 * splats[shBase + n + 3u] + b4 * splats[shBase + n + 4u] + b5 * splats[shBase + n + 5u] + b6 * splats[shBase + n + 6u] + b7 * splats[shBase + n + 7u];
-        cB = cB + b3 * splats[shBase + 2u * n + 3u] + b4 * splats[shBase + 2u * n + 4u] + b5 * splats[shBase + 2u * n + 5u] + b6 * splats[shBase + 2u * n + 6u] + b7 * splats[shBase + 2u * n + 7u];
-    }` : ''}
-    ${coeffsPerChannel >= 15 ? `
-    {
-        let n = COEFFS_PER_CHANNEL;
-        let shBase = base + 14u;
-        let xx2 = dirX * dirX;
-        let yy2 = dirY * dirY;
-        let zz2 = dirZ * dirZ;
-        let xy2 = dirX * dirY;
-        let b8 = SH_C3_0 * dirY * (3.0 * xx2 - yy2);
-        let b9 = SH_C3_1 * xy2 * dirZ;
-        let b10 = SH_C3_2 * dirY * (4.0 * zz2 - xx2 - yy2);
-        let b11 = SH_C3_3 * dirZ * (2.0 * zz2 - 3.0 * xx2 - 3.0 * yy2);
-        let b12 = SH_C3_4 * dirX * (4.0 * zz2 - xx2 - yy2);
-        let b13 = SH_C3_5 * dirZ * (xx2 - yy2);
-        let b14 = SH_C3_6 * dirX * (xx2 - 3.0 * yy2);
-        cR = cR + b8 * splats[shBase + 8u] + b9 * splats[shBase + 9u] + b10 * splats[shBase + 10u] + b11 * splats[shBase + 11u] + b12 * splats[shBase + 12u] + b13 * splats[shBase + 13u] + b14 * splats[shBase + 14u];
-        cG = cG + b8 * splats[shBase + n + 8u] + b9 * splats[shBase + n + 9u] + b10 * splats[shBase + n + 10u] + b11 * splats[shBase + n + 11u] + b12 * splats[shBase + n + 12u] + b13 * splats[shBase + n + 13u] + b14 * splats[shBase + n + 14u];
-        cB = cB + b8 * splats[shBase + 2u * n + 8u] + b9 * splats[shBase + 2u * n + 9u] + b10 * splats[shBase + 2u * n + 10u] + b11 * splats[shBase + 2u * n + 11u] + b12 * splats[shBase + 2u * n + 12u] + b13 * splats[shBase + 2u * n + 13u] + b14 * splats[shBase + 2u * n + 14u];
-    }` : ''}
-
-    let colR = max(0.0, cR + 0.5);
-    let colG = max(0.0, cG + 0.5);
-    let colB = max(0.0, cB + 0.5);
-
-    let alpha = (1.0 / (1.0 + exp(-opacity))) * radiusFade;
-
-    projected[i * 3u + 0u] = vec4<f32>(screenX, screenY, radius, 0.0);
-    projected[i * 3u + 1u] = vec4<f32>(covInvA, covInvB, covInvC, alpha);
-    projected[i * 3u + 2u] = vec4<f32>(colR, colG, colB, 0.0);
-
-    // Per-splat tile-coverage count, clamped at maxCoveragePerSplat.
-    // Tile indices are GROUP-LOCAL (= image-tile-index minus the
-    // group's origin in tiles), so values cover [0, groupTilesX-1] ×
-    // [0, groupTilesY-1] for every group regardless of its position in
-    // the image. Splats outside the group's pixel rectangle were
-    // already culled by the AABB check above.
-    let tsz: f32 = ${wgslF32(TILE_SIZE)};
-    let gox = f32(uniforms.groupPixelOriginX);
-    let goy = f32(uniforms.groupPixelOriginY);
-${projection === 'pinhole' ? /* wgsl */`
-    let minTX = max(0, i32(floor((screenX - radius - gox) / tsz)));
-    let maxTX = min(i32(uniforms.groupTilesX) - 1, i32(floor((screenX + radius - gox) / tsz)));
-    let minTY = max(0, i32(floor((screenY - radius - goy) / tsz)));
-    let maxTY = min(i32(uniforms.groupTilesY) - 1, i32(floor((screenY + radius - goy) / tsz)));
-    if (maxTX < minTX || maxTY < minTY) {
-        coverage[i] = 0u;
-    } else {
-        let raw = u32(maxTX - minTX + 1) * u32(maxTY - minTY + 1);
-        coverage[i] = min(raw, ${maxCoveragePerSplat}u);
-    }
-` : /* wgsl */`
-    // Equirect: the X tile range can extend past the image edges into
-    // negative or > groupTilesX-1 indices — those represent the same
-    // splat seen across the ±π longitude seam. Coverage is the raw
-    // span (capped at groupTilesX so a splat with radius > image_width
-    // doesn't emit duplicate tile keys); emit-pairs walks [minTXraw ..
-    // maxTXraw] in lock-step and applies a modular wrap when writing
-    // tile keys. Y is clamped normally — equirect doesn't wrap across
-    // poles.
-    let minTXraw = i32(floor((screenX - radius - gox) / tsz));
-    let maxTXraw = i32(floor((screenX + radius - gox) / tsz));
-    let txCountRaw = maxTXraw - minTXraw + 1;
-    let txCount = min(txCountRaw, i32(uniforms.groupTilesX));
-    let minTY = max(0, i32(floor((screenY - radius - goy) / tsz)));
-    let maxTY = min(i32(uniforms.groupTilesY) - 1, i32(floor((screenY + radius - goy) / tsz)));
-    if (txCount <= 0 || maxTY < minTY) {
-        coverage[i] = 0u;
-    } else {
-        let raw = u32(txCount) * u32(maxTY - minTY + 1);
-        coverage[i] = min(raw, ${maxCoveragePerSplat}u);
-    }
-`}
-}
-`;
-
-/**
- * Tile-bin emit-pairs shader. For each projected splat, emits
- * `coverage[i]` (tile, splat) pairs into two parallel buffers,
- * starting at `emitOffset[i]`:
- *
- *   tileKeys [emitOffset[i] + j] = tileIdx
- *   splatValues[emitOffset[i] + j] = splatIdx (= i)
- *
- * The orchestrator sizes maxCoveragePerSplat to cover a sub-frame's
- * entire tile area, so a splat's `coverage[i]` always equals its full
- * bbox-in-group tile count — no truncation, no seams. The walk emits
- * row-major over the bbox-in-group. A subsequent key+value radix sort
- * groups pairs by tile; within each tile, the splatIdx-as-value sort
- * preserves the chunk's depth order (splatIdx is monotonic in depth
- * from the CPU pre-sort).
- *
- * @param projection - Camera projection. `pinhole` walks the clamped
- * tile bbox directly; `equirect` walks the un-clamped X range and
- * applies a modular wrap, so a splat near the ±π longitude seam emits
- * tile keys on both sides of the image.
- * @returns WGSL source for the emit-pairs compute shader.
- */
-const tileBinEmitPairsWgsl = (projection: Projection) => /* wgsl */`
-struct Uniforms {
-    rightX: f32, rightY: f32, rightZ: f32, _p0: f32,
-    downX: f32, downY: f32, downZ: f32, _p1: f32,
-    forwardX: f32, forwardY: f32, forwardZ: f32, _p2: f32,
-    eyeX: f32, eyeY: f32, eyeZ: f32, _p3: f32,
-    focalX: f32, focalY: f32, near: f32, _p4: f32,
-    imageWidth: u32, imageHeight: u32, splatStride: u32, chunkSize: u32,
-    groupPixelMinX: u32, groupPixelMinY: u32, groupPixelMaxX: u32, groupPixelMaxY: u32,
-    groupTilesX: u32, groupTilesY: u32, groupPixelOriginX: u32, groupPixelOriginY: u32,
-    bgR: f32, bgG: f32, bgB: f32, bgA: f32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: Uniforms;
-@group(0) @binding(1) var<storage, read> projected: array<vec4<f32>>;
-@group(0) @binding(2) var<storage, read> emitOffset: array<u32>;
-@group(0) @binding(3) var<storage, read> coverage: array<u32>;
-@group(0) @binding(4) var<storage, read_write> tileKeys: array<u32>;
-@group(0) @binding(5) var<storage, read_write> splatValues: array<u32>;
-
-@compute @workgroup_size(64)
-fn main(@builtin(global_invocation_id) gid: vec3<u32>) {
-    let i = gid.x;
-    if (i >= uniforms.chunkSize) { return; }
-    let cap = coverage[i];
-    if (cap == 0u) { return; }
-    let v0 = projected[i * 3u + 0u];
-    let radius = v0.z;
-    if (radius <= 0.0) { return; }
-    let sX = v0.x;
-    let sY = v0.y;
-    let tsz: f32 = ${wgslF32(TILE_SIZE)};
-    // Group-local tile indices (see project shader for rationale).
-    let gox = f32(uniforms.groupPixelOriginX);
-    let goy = f32(uniforms.groupPixelOriginY);
-${projection === 'pinhole' ? /* wgsl */`
-    let minTX = max(0, i32(floor((sX - radius - gox) / tsz)));
-    let maxTX = min(i32(uniforms.groupTilesX) - 1, i32(floor((sX + radius - gox) / tsz)));
-    let minTY = max(0, i32(floor((sY - radius - goy) / tsz)));
-    let maxTY = min(i32(uniforms.groupTilesY) - 1, i32(floor((sY + radius - goy) / tsz)));
-    if (maxTX < minTX || maxTY < minTY) { return; }
-
-    var slot = emitOffset[i];
-    let end = slot + cap;
-    for (var ty: i32 = minTY; ty <= maxTY; ty = ty + 1) {
-        if (slot >= end) { break; }
-        for (var tx: i32 = minTX; tx <= maxTX; tx = tx + 1) {
-            if (slot >= end) { break; }
-            let t = u32(ty) * uniforms.groupTilesX + u32(tx);
-            tileKeys[slot] = t;
-            splatValues[slot] = i;
-            slot = slot + 1u;
-        }
-    }
-` : /* wgsl */`
-    // Equirect: raw X range (possibly wrapping past the seam) — must
-    // match the project shader's coverage computation exactly. Each
-    // emitted tx is wrapped into [0, groupTilesX-1] via modular
-    // arithmetic. The rasterize-binned shader compensates by wrapping
-    // its per-pixel dx into [-W/2, W/2], so a wrapped tile pulls the
-    // splat's footprint from the correct copy across the seam.
-    let minTXraw = i32(floor((sX - radius - gox) / tsz));
-    let maxTXraw = i32(floor((sX + radius - gox) / tsz));
-    let txCountRaw = maxTXraw - minTXraw + 1;
-    let groupTilesX_i = i32(uniforms.groupTilesX);
-    let txCount = min(txCountRaw, groupTilesX_i);
-    let minTY = max(0, i32(floor((sY - radius - goy) / tsz)));
-    let maxTY = min(i32(uniforms.groupTilesY) - 1, i32(floor((sY + radius - goy) / tsz)));
-    if (txCount <= 0 || maxTY < minTY) { return; }
-
-    var slot = emitOffset[i];
-    let end = slot + cap;
-    for (var ty: i32 = minTY; ty <= maxTY; ty = ty + 1) {
-        if (slot >= end) { break; }
-        for (var k: i32 = 0; k < txCount; k = k + 1) {
-            if (slot >= end) { break; }
-            var tx = (minTXraw + k) % groupTilesX_i;
-            if (tx < 0) { tx = tx + groupTilesX_i; }
-            let t = u32(ty) * uniforms.groupTilesX + u32(tx);
-            tileKeys[slot] = t;
-            splatValues[slot] = i;
-            slot = slot + 1u;
-        }
-    }
-`}
-}
-`;
-
-/**
- * Single-workgroup exclusive prefix-sum of the project shader's
- * per-splat `coverage[]` into `emitOffset[]`. Also writes the total
- * pair count into `totalPairs[0]` so downstream kernels and the radix
- * sort can size their dispatches without a CPU round trip.
- *
- * Layout: 256 threads, each processes `${SCAN_PER_THREAD}` elements
- * serially (chosen so that 256 × SCAN_PER_THREAD ≥ chunkCap). Phase 1
- * computes a per-thread partial sum; phase 2 has thread 0 scan the 256
- * partials in shared memory (negligible vs the 800-element block work);
- * phase 3 each thread re-walks its block writing the exclusive prefix.
- *
- * @param scanPerThread - Per-thread element budget; must satisfy `256 * scanPerThread >= chunkCap`.
- * @returns WGSL source for the prefix-sum compute shader.
- */
-const prefixSumWgsl = (scanPerThread: number) => /* wgsl */`
-struct Uniforms {
-    rightX: f32, rightY: f32, rightZ: f32, _p0: f32,
-    downX: f32, downY: f32, downZ: f32, _p1: f32,
-    forwardX: f32, forwardY: f32, forwardZ: f32, _p2: f32,
-    eyeX: f32, eyeY: f32, eyeZ: f32, _p3: f32,
-    focalX: f32, focalY: f32, near: f32, _p4: f32,
-    imageWidth: u32, imageHeight: u32, splatStride: u32, chunkSize: u32,
-    groupPixelMinX: u32, groupPixelMinY: u32, groupPixelMaxX: u32, groupPixelMaxY: u32,
-    groupTilesX: u32, groupTilesY: u32, groupPixelOriginX: u32, groupPixelOriginY: u32,
-    bgR: f32, bgG: f32, bgB: f32, bgA: f32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: Uniforms;
-@group(0) @binding(1) var<storage, read> coverage: array<u32>;
-@group(0) @binding(2) var<storage, read_write> emitOffset: array<u32>;
-@group(0) @binding(3) var<storage, read_write> totalPairs: array<u32>;
-
-const SCAN_THREADS: u32 = 256u;
-const SCAN_PER_THREAD: u32 = ${scanPerThread}u;
-
-var<workgroup> scratch: array<u32, SCAN_THREADS>;
-
-@compute @workgroup_size(SCAN_THREADS)
-fn main(@builtin(local_invocation_id) lid: vec3<u32>) {
-    let tid = lid.x;
-    let n = uniforms.chunkSize;
-    let base = tid * SCAN_PER_THREAD;
-
-    var partial: u32 = 0u;
-    for (var i: u32 = 0u; i < SCAN_PER_THREAD; i = i + 1u) {
-        let idx = base + i;
-        if (idx < n) {
-            partial = partial + coverage[idx];
-        }
-    }
-    scratch[tid] = partial;
-    workgroupBarrier();
-
-    if (tid == 0u) {
-        var acc: u32 = 0u;
-        for (var i: u32 = 0u; i < SCAN_THREADS; i = i + 1u) {
-            let v = scratch[i];
-            scratch[i] = acc;
-            acc = acc + v;
-        }
-        totalPairs[0] = acc;
-    }
-    workgroupBarrier();
-
-    var prefix: u32 = scratch[tid];
-    for (var i: u32 = 0u; i < SCAN_PER_THREAD; i = i + 1u) {
-        let idx = base + i;
-        if (idx < n) {
-            emitOffset[idx] = prefix;
-            prefix = prefix + coverage[idx];
-        }
-    }
-}
-`;
-
-/**
- * Prepares indirect-dispatch arguments by reading `totalPairs[0]` and
- * writing workgroup counts into two slots of the device's
- * `indirectDispatchBuffer` (each slot is 3 × u32 = `(x, y, z)`):
- *
- *   - `sortSlot`: workgroup count for `ComputeRadixSort.sortIndirect`,
- *     computed as `ceil(totalPairs / 2048)` (matches the radix sort's
- *     16×16 thread × 8 elements / thread = 2048 elements/workgroup).
- *   - `boundariesSlot`: workgroup count for `findBoundaries`, computed
- *     as `ceil(totalPairs / 64)`.
- *
- * Slot byte offsets are passed in via two u32 uniforms in a small ad-hoc
- * uniform block (NOT the shared `Uniforms` struct, because the slot
- * indices vary per chunk while the shared uniforms are set per group).
- *
- * @returns WGSL source for the prepare-indirect compute shader.
- */
-const prepareIndirectWgsl = () => /* wgsl */`
-struct PrepareUniforms {
-    sortSlotBase: u32,
-    boundariesSlotBase: u32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: PrepareUniforms;
-@group(0) @binding(1) var<storage, read> totalPairs: array<u32>;
-@group(0) @binding(2) var<storage, read_write> indirectBuffer: array<u32>;
-
-const SORT_ELEMENTS_PER_WG: u32 = 2048u;
-const BOUNDARIES_THREADS_PER_WG: u32 = 64u;
-// WebGPU spec minimum for maxComputeWorkgroupsPerDimension. Any larger
-// 1-D dispatch must be tiled into 2-D so both axes stay <= this bound.
-// The consumer shaders linearise via WORKGROUP_ID = w_id.x + w_id.y * w_dim.x.
-const MAX_DIM: u32 = 65535u;
-
-fn splitWg(count: u32) -> vec2<u32> {
-    if (count <= MAX_DIM) {
-        return vec2<u32>(count, 1u);
-    }
-    let y = (count + MAX_DIM - 1u) / MAX_DIM;
-    let x = (count + y - 1u) / y;
-    return vec2<u32>(x, y);
-}
-
-@compute @workgroup_size(1)
-fn main() {
-    let n = totalPairs[0];
-    let sortWg = (n + SORT_ELEMENTS_PER_WG - 1u) / SORT_ELEMENTS_PER_WG;
-    let bndWg = (n + BOUNDARIES_THREADS_PER_WG - 1u) / BOUNDARIES_THREADS_PER_WG;
-    let sortDim = splitWg(sortWg);
-    let bndDim = splitWg(bndWg);
-    let s = uniforms.sortSlotBase;
-    let b = uniforms.boundariesSlotBase;
-    indirectBuffer[s + 0u] = sortDim.x;
-    indirectBuffer[s + 1u] = sortDim.y;
-    indirectBuffer[s + 2u] = 1u;
-    indirectBuffer[b + 0u] = bndDim.x;
-    indirectBuffer[b + 1u] = bndDim.y;
-    indirectBuffer[b + 2u] = 1u;
-}
-`;
-
-/**
- * Initialises `tileOffsets[0 .. numTiles]` to the sentinel value
- * `totalPairs[0]` (= past-the-end). `findBoundaries` then atomicMin's
- * the actual first-pair-index for every non-empty tile; tiles with no
- * pairs keep the sentinel, which collapses to a zero-length slice when
- * the rasterize-binned shader reads `tileOffsets[T] .. tileOffsets[T+1]`.
- *
- * @returns WGSL source for the init-tile-offsets compute shader.
- */
-const initTileOffsetsWgsl = () => /* wgsl */`
-struct Uniforms {
-    rightX: f32, rightY: f32, rightZ: f32, _p0: f32,
-    downX: f32, downY: f32, downZ: f32, _p1: f32,
-    forwardX: f32, forwardY: f32, forwardZ: f32, _p2: f32,
-    eyeX: f32, eyeY: f32, eyeZ: f32, _p3: f32,
-    focalX: f32, focalY: f32, near: f32, _p4: f32,
-    imageWidth: u32, imageHeight: u32, splatStride: u32, chunkSize: u32,
-    groupPixelMinX: u32, groupPixelMinY: u32, groupPixelMaxX: u32, groupPixelMaxY: u32,
-    groupTilesX: u32, groupTilesY: u32, groupPixelOriginX: u32, groupPixelOriginY: u32,
-    bgR: f32, bgG: f32, bgB: f32, bgA: f32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: Uniforms;
-@group(0) @binding(1) var<storage, read> totalPairs: array<u32>;
-@group(0) @binding(2) var<storage, read_write> tileOffsets: array<u32>;
-
-@compute @workgroup_size(64)
-fn main(@builtin(global_invocation_id) gid: vec3<u32>) {
-    let i = gid.x;
-    let numTiles = uniforms.groupTilesX * uniforms.groupTilesY;
-    if (i > numTiles) { return; }
-    tileOffsets[i] = totalPairs[0];
-}
-`;
-
-/**
- * For every adjacent pair of sorted keys where the high-bit tile index
- * differs, atomicMin's the current position into `tileOffsets[t]` for
- * every tile `t` in `(prevTile, curTile]`. Combined with the sentinel
- * init this gives `tileOffsets[T]` = first index in `sortedKeys` whose
- * tile bits equal T (or the sentinel if T is empty).
- *
- * Dispatched indirectly with workgroup count `ceil(totalPairs / 64)` so
- * that we don't waste invocations on the unused tail of the pairs
- * buffer.
- *
- * @returns WGSL source for the find-boundaries compute shader.
- */
-const findBoundariesWgsl = () => /* wgsl */`
-struct Uniforms {
-    rightX: f32, rightY: f32, rightZ: f32, _p0: f32,
-    downX: f32, downY: f32, downZ: f32, _p1: f32,
-    forwardX: f32, forwardY: f32, forwardZ: f32, _p2: f32,
-    eyeX: f32, eyeY: f32, eyeZ: f32, _p3: f32,
-    focalX: f32, focalY: f32, near: f32, _p4: f32,
-    imageWidth: u32, imageHeight: u32, splatStride: u32, chunkSize: u32,
-    groupPixelMinX: u32, groupPixelMinY: u32, groupPixelMaxX: u32, groupPixelMaxY: u32,
-    groupTilesX: u32, groupTilesY: u32, groupPixelOriginX: u32, groupPixelOriginY: u32,
-    bgR: f32, bgG: f32, bgB: f32, bgA: f32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: Uniforms;
-@group(0) @binding(1) var<storage, read> totalPairs: array<u32>;
-@group(0) @binding(2) var<storage, read> sortedTileKeys: array<u32>;
-@group(0) @binding(3) var<storage, read_write> tileOffsets: array<atomic<u32>>;
-
-@compute @workgroup_size(64)
-fn main(
-    @builtin(workgroup_id) wgId: vec3<u32>,
-    @builtin(num_workgroups) numWg: vec3<u32>,
-    @builtin(local_invocation_id) lid: vec3<u32>
-) {
-    // 2-D dispatch (prepare-indirect splits to stay under the 65535
-    // per-axis workgroup-count limit); linearise here.
-    let linearWg = wgId.x + wgId.y * numWg.x;
-    let i = linearWg * 64u + lid.x;
-    let n = totalPairs[0];
-    if (i >= n) { return; }
-
-    // Reference uniforms once so the binding isn't dead-code-stripped
-    // (keeps the BG format in sync with the shader expectations).
-    let _u = uniforms.groupTilesX;
-
-    let curTile = sortedTileKeys[i];
-    // Sentinel for "no previous tile" — overflow makes prevTile+1 = 0u
-    // so the for loop below cleanly handles the i = 0 case.
-    let prevTileBits = select(0xFFFFFFFFu, sortedTileKeys[i - 1u], i > 0u);
-    if (curTile == prevTileBits) { return; }
-    for (var t: u32 = prevTileBits + 1u; t <= curTile; t = t + 1u) {
-        atomicMin(&tileOffsets[t], i);
-    }
-}
-`;
-
-/**
- * Binned rasterize shader. Each workgroup handles one tile and only walks
- * the splats that have been pre-binned into it (tile-bin pre-pass on CPU
- * or GPU). Replaces the "walk all splats per pixel" loop in
- * `rasterizeWgsl` with "walk this tile's slice", which is the asymptotic
- * fix for performance at high splat counts.
- *
- * The slice is stored in two buffers:
- *   - `tileOffsets[T + 1]` — exclusive prefix sum: tile T's slice is
- *     `tileData[tileOffsets[T] .. tileOffsets[T + 1])`.
- *   - `tileData[]` — splat indices, grouped by tile, depth-sorted within
- *     each tile (the orchestrator's CPU pre-sort + stable per-splat
- *     binning produces this layout for free).
- *
- * The shader's bindings mirror `rasterizeWgsl` and add bindings 3, 4 for
- * the tile lists.
- *
- * @param projection - Camera projection. `equirect` wraps the per-pixel
- * `dx = px - splat.x` into `[-W/2, W/2]` so a tile on the opposite side
- * of the ±π longitude seam evaluates against the splat's nearer copy;
- * `pinhole` uses the raw delta.
- * @returns WGSL source for the binned-rasterize compute shader.
- */
-const rasterizeBinnedWgsl = (projection: Projection) => /* wgsl */`
-struct Uniforms {
-    rightX: f32, rightY: f32, rightZ: f32, _p0: f32,
-    downX: f32, downY: f32, downZ: f32, _p1: f32,
-    forwardX: f32, forwardY: f32, forwardZ: f32, _p2: f32,
-    eyeX: f32, eyeY: f32, eyeZ: f32, _p3: f32,
-    focalX: f32, focalY: f32, near: f32, _p4: f32,
-    imageWidth: u32, imageHeight: u32, splatStride: u32, chunkSize: u32,
-    groupPixelMinX: u32, groupPixelMinY: u32, groupPixelMaxX: u32, groupPixelMaxY: u32,
-    groupTilesX: u32, groupTilesY: u32, groupPixelOriginX: u32, groupPixelOriginY: u32,
-    bgR: f32, bgG: f32, bgB: f32, bgA: f32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: Uniforms;
-@group(0) @binding(1) var<storage, read> projected: array<vec4<f32>>;
-@group(0) @binding(2) var<storage, read_write> runningState: array<vec4<f32>>;
-@group(0) @binding(3) var<storage, read> tileOffsets: array<u32>;
-@group(0) @binding(4) var<storage, read> sortedSplatIndices: array<u32>;
-
-@compute @workgroup_size(${TILE_SIZE}, ${TILE_SIZE}, 1)
-fn main(
-    @builtin(workgroup_id) wgId: vec3<u32>,
-    @builtin(local_invocation_id) lid: vec3<u32>
-) {
-    if (wgId.x >= uniforms.groupTilesX || wgId.y >= uniforms.groupTilesY) { return; }
-
-    let tileIdx = wgId.y * uniforms.groupTilesX + wgId.x;
-    let sliceStart = tileOffsets[tileIdx];
-    let sliceEnd = tileOffsets[tileIdx + 1u];
-
-    let localPixelX = wgId.x * ${TILE_SIZE}u + lid.x;
-    let localPixelY = wgId.y * ${TILE_SIZE}u + lid.y;
-    let groupPixelW = uniforms.groupTilesX * ${TILE_SIZE}u;
-
-    let imagePixelX = uniforms.groupPixelOriginX + localPixelX;
-    let imagePixelY = uniforms.groupPixelOriginY + localPixelY;
-    if (imagePixelX >= uniforms.imageWidth || imagePixelY >= uniforms.imageHeight) { return; }
-
-    let pixelIdx = localPixelY * groupPixelW + localPixelX;
-    var state = runningState[pixelIdx];
-    var color = state.rgb;
-    var T = state.a;
-
-    if (T < ${wgslF32(MIN_TRANSMITTANCE)}) { return; }
-
-    let px = f32(imagePixelX) + 0.5;
-    let py = f32(imagePixelY) + 0.5;
-
-${projection === 'equirect' ? /* wgsl */`
-    let imgWf2 = f32(uniforms.imageWidth);
-    let halfImgW = imgWf2 * 0.5;
-` : ''}
-    for (var i: u32 = sliceStart; i < sliceEnd; i = i + 1u) {
-        if (T < ${wgslF32(MIN_TRANSMITTANCE)}) { break; }
-        let splatIdx = sortedSplatIndices[i];
-        let v0 = projected[splatIdx * 3u + 0u];
-${projection === 'pinhole' ? /* wgsl */`
-        let dx = px - v0.x;
-` : /* wgsl */`
-        // Equirect: a splat near the ±π longitude seam is tile-binned on
-        // both sides of the image. Wrap dx into [-W/2, W/2] so a tile on
-        // the opposite side of the seam pulls the splat's footprint from
-        // the correct (nearer) copy.
-        var dx = px - v0.x;
-        if (dx > halfImgW) { dx = dx - imgWf2; }
-        else if (dx < -halfImgW) { dx = dx + imgWf2; }
-`}
-        let dy = py - v0.y;
-        let r = v0.z;
-        if (r <= 0.0 || abs(dx) > r || abs(dy) > r) { continue; }
-        let v1 = projected[splatIdx * 3u + 1u];
-        let power = -0.5 * (v1.x * dx * dx + 2.0 * v1.y * dx * dy + v1.z * dy * dy);
-        if (power > 0.0) { continue; }
-        let alpha = min(${wgslF32(OPACITY_CAP)}, v1.w * exp(power));
-        if (alpha < ${wgslF32(MIN_ALPHA)}) { continue; }
-        let weight = T * alpha;
-        let v2 = projected[splatIdx * 3u + 2u];
-        color = color + weight * v2.rgb;
-        T = T * (1.0 - alpha);
-    }
-
-    runningState[pixelIdx] = vec4<f32>(color, T);
-}
-`;
-
-const finalizeWgsl = () => /* wgsl */`
-struct Uniforms {
-    rightX: f32, rightY: f32, rightZ: f32, _p0: f32,
-    downX: f32, downY: f32, downZ: f32, _p1: f32,
-    forwardX: f32, forwardY: f32, forwardZ: f32, _p2: f32,
-    eyeX: f32, eyeY: f32, eyeZ: f32, _p3: f32,
-    focalX: f32, focalY: f32, near: f32, _p4: f32,
-    imageWidth: u32, imageHeight: u32, splatStride: u32, chunkSize: u32,
-    groupPixelMinX: u32, groupPixelMinY: u32, groupPixelMaxX: u32, groupPixelMaxY: u32,
-    groupTilesX: u32, groupTilesY: u32, groupPixelOriginX: u32, groupPixelOriginY: u32,
-    bgR: f32, bgG: f32, bgB: f32, bgA: f32,
-}
-
-@group(0) @binding(0) var<uniform> uniforms: Uniforms;
-@group(0) @binding(1) var<storage, read> runningState: array<vec4<f32>>;
-@group(0) @binding(2) var<storage, read_write> output: array<u32>;
-
-@compute @workgroup_size(${TILE_SIZE}, ${TILE_SIZE}, 1)
-fn main(
-    @builtin(workgroup_id) wgId: vec3<u32>,
-    @builtin(local_invocation_id) lid: vec3<u32>
-) {
-    if (wgId.x >= uniforms.groupTilesX || wgId.y >= uniforms.groupTilesY) { return; }
-
-    let localPixelX = wgId.x * ${TILE_SIZE}u + lid.x;
-    let localPixelY = wgId.y * ${TILE_SIZE}u + lid.y;
-    let groupPixelW = uniforms.groupTilesX * ${TILE_SIZE}u;
-
-    let pixelIdx = localPixelY * groupPixelW + localPixelX;
-    let state = runningState[pixelIdx];
-
-    let color = state.rgb + state.a * vec3<f32>(uniforms.bgR, uniforms.bgG, uniforms.bgB);
-    let alphaOut = (1.0 - state.a) + state.a * uniforms.bgA;
-
-    let r = u32(clamp(color.r, 0.0, 1.0) * 255.0 + 0.5);
-    let g = u32(clamp(color.g, 0.0, 1.0) * 255.0 + 0.5);
-    let bch = u32(clamp(color.b, 0.0, 1.0) * 255.0 + 0.5);
-    let aOut = u32(clamp(alphaOut, 0.0, 1.0) * 255.0 + 0.5);
-
-    output[pixelIdx] = r | (g << 8u) | (bch << 16u) | (aOut << 24u);
-}
-`;
-
-/**
- * Build the UniformFormat entries that match the WGSL uniform block above.
- * All three shaders share the same uniform layout so a single
- * UniformBufferFormat description suffices (we instantiate it per shader).
- *
- * @returns Array of UniformFormat entries in declaration order.
- */
-const uniformFormatEntries = (): UniformFormat[] => [
-    new UniformFormat('rightX', UNIFORMTYPE_FLOAT),
-    new UniformFormat('rightY', UNIFORMTYPE_FLOAT),
-    new UniformFormat('rightZ', UNIFORMTYPE_FLOAT),
-    new UniformFormat('_p0', UNIFORMTYPE_FLOAT),
-    new UniformFormat('downX', UNIFORMTYPE_FLOAT),
-    new UniformFormat('downY', UNIFORMTYPE_FLOAT),
-    new UniformFormat('downZ', UNIFORMTYPE_FLOAT),
-    new UniformFormat('_p1', UNIFORMTYPE_FLOAT),
-    new UniformFormat('forwardX', UNIFORMTYPE_FLOAT),
-    new UniformFormat('forwardY', UNIFORMTYPE_FLOAT),
-    new UniformFormat('forwardZ', UNIFORMTYPE_FLOAT),
-    new UniformFormat('_p2', UNIFORMTYPE_FLOAT),
-    new UniformFormat('eyeX', UNIFORMTYPE_FLOAT),
-    new UniformFormat('eyeY', UNIFORMTYPE_FLOAT),
-    new UniformFormat('eyeZ', UNIFORMTYPE_FLOAT),
-    new UniformFormat('_p3', UNIFORMTYPE_FLOAT),
-    new UniformFormat('focalX', UNIFORMTYPE_FLOAT),
-    new UniformFormat('focalY', UNIFORMTYPE_FLOAT),
-    new UniformFormat('near', UNIFORMTYPE_FLOAT),
-    new UniformFormat('_p4', UNIFORMTYPE_FLOAT),
-    new UniformFormat('imageWidth', UNIFORMTYPE_UINT),
-    new UniformFormat('imageHeight', UNIFORMTYPE_UINT),
-    new UniformFormat('splatStride', UNIFORMTYPE_UINT),
-    new UniformFormat('chunkSize', UNIFORMTYPE_UINT),
-    new UniformFormat('groupPixelMinX', UNIFORMTYPE_UINT),
-    new UniformFormat('groupPixelMinY', UNIFORMTYPE_UINT),
-    new UniformFormat('groupPixelMaxX', UNIFORMTYPE_UINT),
-    new UniformFormat('groupPixelMaxY', UNIFORMTYPE_UINT),
-    new UniformFormat('groupTilesX', UNIFORMTYPE_UINT),
-    new UniformFormat('groupTilesY', UNIFORMTYPE_UINT),
-    new UniformFormat('groupPixelOriginX', UNIFORMTYPE_UINT),
-    new UniformFormat('groupPixelOriginY', UNIFORMTYPE_UINT),
-    new UniformFormat('bgR', UNIFORMTYPE_FLOAT),
-    new UniformFormat('bgG', UNIFORMTYPE_FLOAT),
-    new UniformFormat('bgB', UNIFORMTYPE_FLOAT),
-    new UniformFormat('bgA', UNIFORMTYPE_FLOAT)
-];
 
 interface PipelineBuffers {
     inputBuffer: StorageBuffer;
@@ -1250,21 +303,69 @@ class GpuSplatRasterizer {
             new BindStorageBufferFormat('output', SHADERSTAGE_COMPUTE)
         ]);
 
+        // WGSL chunks made available to every shader. The engine's
+        // compute-shader path runs the source through
+        // `Preprocessor.run(cshader, cincludes, ...)`, so any
+        // `#include "name"` directive resolves to the corresponding
+        // value in this Map. Lookup is by name only — chunk-order
+        // does not matter.
+        //
+        // Per-render-cap chunks (tileAabb / tileWalk) are constructed
+        // here with the resolved `maxCoveragePerSplat` so the chunk
+        // bodies stay JS-template-free.
+        const projection = options.projection;
+        const sharedCincludes = new Map<string, string>([
+            ['uniformsStruct', uniformsStruct],
+            ['constants', constantsChunk],
+            ['projectionPinhole', projectionPinhole],
+            ['projectionEquirect', projectionEquirect],
+            ['jacobianPinhole', jacobianPinhole],
+            ['jacobianEquirect', jacobianEquirect],
+            ['tileAabbPinhole', tileAabbPinhole(options.maxCoveragePerSplat)],
+            ['tileAabbEquirect', tileAabbEquirect(options.maxCoveragePerSplat)],
+            ['tileWalkPinhole', tileWalkPinhole],
+            ['tileWalkEquirect', tileWalkEquirect],
+            ['shBand1', shBand1],
+            ['shBand2', shBand2],
+            ['shBand3', shBand3],
+            ['quatRotation', quatRotation],
+            ['covariance3D', covariance3D]
+        ]);
+
+        // Per-render variant flags consumed by `#ifdef` directives in
+        // the WGSL sources. Presence-only — the empty value is fine
+        // because the preprocessor only checks `defines.has(name)` for
+        // `#ifdef`. See `engine/src/core/preprocessor.js` for how
+        // `cdefines` is consumed.
+        const sharedCdefines = new Map<string, string>();
+        if (projection === 'equirect') {
+            sharedCdefines.set('PROJECTION_EQUIRECT', '');
+        }
+        if (options.numSHBands >= 1) sharedCdefines.set('SH_BAND_1', '');
+        if (options.numSHBands >= 2) sharedCdefines.set('SH_BAND_2', '');
+        if (options.numSHBands >= 3) sharedCdefines.set('SH_BAND_3', '');
+
         const mkShader = (
             name: string,
             source: string,
             bgFormat: BindGroupFormat,
-            uniformEntries: UniformFormat[] = uniformFormatEntries()
+            uniformEntries: UniformFormat[] = uniformFormatEntries(),
+            cincludes: Map<string, string> = sharedCincludes,
+            cdefines: Map<string, string> = sharedCdefines
         ) => new Shader(device, {
             name,
             shaderLanguage: SHADERLANGUAGE_WGSL,
             cshader: source,
-            // @ts-ignore - computeUniformBufferFormats / computeBindGroupFormat are not in public Shader types.
+            // @ts-ignore - computeUniformBufferFormats / computeBindGroupFormat / cincludes / cdefines are not in public Shader types.
             computeUniformBufferFormats: {
                 uniforms: new UniformBufferFormat(device, uniformEntries)
             },
             // @ts-ignore
-            computeBindGroupFormat: bgFormat
+            computeBindGroupFormat: bgFormat,
+            // @ts-ignore
+            cincludes,
+            // @ts-ignore
+            cdefines
         });
 
         // The prefix-sum kernel processes the chunk in 256-thread blocks
@@ -1281,14 +382,13 @@ class GpuSplatRasterizer {
             new UniformFormat('boundariesSlotBase', UNIFORMTYPE_UINT)
         ];
 
-        const projection = options.projection;
-        this.projectShader = mkShader('splat-project', projectWgsl(coeffs, options.maxCoveragePerSplat, projection), this.projectBgFormat);
+        this.projectShader = mkShader('splat-project', projectWgsl(coeffs), this.projectBgFormat);
         this.prefixSumShader = mkShader('splat-tilebin-prefix-sum', prefixSumWgsl(scanPerThread), this.prefixSumBgFormat);
-        this.emitPairsShader = mkShader('splat-tilebin-emit-pairs', tileBinEmitPairsWgsl(projection), this.emitPairsBgFormat);
+        this.emitPairsShader = mkShader('splat-tilebin-emit-pairs', tileBinEmitPairsWgsl(), this.emitPairsBgFormat);
         this.prepareIndirectShader = mkShader('splat-tilebin-prepare-indirect', prepareIndirectWgsl(), this.prepareIndirectBgFormat, prepareIndirectUniforms);
         this.initTileOffsetsShader = mkShader('splat-tilebin-init-tile-offsets', initTileOffsetsWgsl(), this.initTileOffsetsBgFormat);
         this.findBoundariesShader = mkShader('splat-tilebin-find-boundaries', findBoundariesWgsl(), this.findBoundariesBgFormat);
-        this.rasterizeBinnedShader = mkShader('splat-rasterize-binned', rasterizeBinnedWgsl(projection), this.rasterizeBinnedBgFormat);
+        this.rasterizeBinnedShader = mkShader('splat-rasterize-binned', rasterizeBinnedWgsl(), this.rasterizeBinnedBgFormat);
         this.finalizeShader = mkShader('splat-finalize-pack', finalizeWgsl(), this.finalizeBgFormat);
 
         // Buffer sizing. runningState/output cover exactly the group's
